ai-service/requirements.txt
streamlit
google-generativeai
python-dotenv
requests
PyPDF2
python-docx